requirements.txt
google-auth>=2.22.0
google-cloud-private-ca>=1.7.0
flask>=2.3.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
flask-cors>=5.0.0
flask-session>=0.5.0
redis>=5.0.0 